remove point in tx

+++ common/common.go
@@ -10,7 +10,7 @@ import (
 	"time"
 )
 
-var Version uint64 = 1
+var Version int32 = 1
 
 func HmAC(message, key []byte) []byte {
 	h := hmac.New(sha512.New, key)
